Remove border from .card
According to the exercise, there should be 8px between the edge of
the card and its content. Initially this was done with an 8px wide
border when it lies between the edge of the card and its margin(s).

Padding is the correct property to use. Visually it makes no
difference but may cause issues in more complex exercises.